Extend the subscription model with codeflow metadata
