+++ src/Maestro/Maestro.Data/Models/Subscription.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
+using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using Microsoft.DotNet.DarcLib;
@@ -67,6 +68,17 @@ public class Subscription
 
     public bool Enabled { get; set; } = true;
 
+    /// <summary>
+    /// Denotes whether sources are also synchronized.
+    /// Source or target repository must be a VMR.
+    /// </summary>
+    public bool SourceEnabled { get; set; }
+
+    /// <summary>
+    /// Dependencies to ignore when synchronizing code of source-enabled subscriptions.
+    /// </summary>
+    public List<AssetFilter> ExcludedDependencies { get; set; }
+
     [NotMapped]
     public SubscriptionPolicy PolicyObject
     {
@@ -79,69 +91,4 @@ public class Subscription
     public Build LastAppliedBuild { get; set; }
 
     public string PullRequestFailureNotificationTags { get; set; }
-}
-
-public class SubscriptionUpdate
-{
-    [Key]
-    public Guid SubscriptionId { get; set; }
-
-    public Subscription Subscription { get; set; }
-
-    /// <summary>
-    ///     **true** if the update succeeded; **false** otherwise.
-    /// </summary>
-    public bool Success { get; set; }
-
-    /// <summary>
-    ///     A message describing what the subscription was trying to do.
-    ///     e.g. 'Updating dependencies from dotnet/coreclr in dotnet/corefx'
-    /// </summary>
-    public string Action { get; set; }
-
-    /// <summary>
-    ///     The error that occured, if any.
-    /// </summary>
-    public string ErrorMessage { get; set; }
-
-    /// <summary>
-    ///     The method that was called.
-    /// </summary>
-    public string Method { get; set; }
-
-    /// <summary>
-    ///     The parameters to the called method.
-    /// </summary>
-    public string Arguments { get; set; }
-}
-
-public class SubscriptionUpdateHistory
-{
-    public Guid SubscriptionId { get; set; }
-
-    /// <summary>
-    ///     **true** if the update succeeded; **false** otherwise.
-    /// </summary>
-    public bool Success { get; set; }
-
-    /// <summary>
-    ///     A message describing what the subscription was trying to do.
-    ///     e.g. 'Updating dependencies from dotnet/coreclr in dotnet/corefx'
-    /// </summary>
-    public string Action { get; set; }
-
-    /// <summary>
-    ///     The error that occured, if any.
-    /// </summary>
-    public string ErrorMessage { get; set; }
-
-    /// <summary>
-    ///     The method that was called.
-    /// </summary>
-    public string Method { get; set; }
-
-    /// <summary>
-    ///     The parameters to the called method.
-    /// </summary>
-    public string Arguments { get; set; }
 }
